Move AllocatedBuffer into its own file
